refactor: extract Markdown parser into lib/MdParser
Per znelson's review suggestion — separate parsing/layout logic from
the reader activity.

- New lib/MdParser/MdParser.h + MdParser.cpp:
  - MdLine struct (text, style, indentPixels, isHRule)
  - MD_BULLET_INDENT constant
  - MdParser::trimWhitespace()
  - MdParser::isHorizontalRule()
  - MdParser::stripInlineMarkdown()
  - MdParser::parseMdLine()
  Zero dependencies beyond EpdFont (no renderer, no settings, no SD).

- MdReaderActivity.h: remove MdLine struct definition; include MdParser.h
- MdReaderActivity.cpp: remove ~200 lines of static helpers; use
  MdParser:: namespace for all calls

Activity file: 902 → 703 lines. Parser is now independently reusable
and unit-testable (future search/index features can include MdParser.h
without pulling in the reader activity).
